+++ packages/create-theme/public/template/docs/.vitepress/blog-theme.ts
@@ -4,15 +4,15 @@ import { getThemeConfig } from '@sugarat/theme/node'
 // 开启RSS支持（RSS配置）
 // import type { Theme } from '@sugarat/theme'
 
-// const baseUrl = 'https://www.dmsrs.org'
+// const baseUrl = 'https://sugarat.top'
 // const RSS: Theme.RSSOptions = {
-//   title: '代码收容所',
+//   title: '粥里有勺糖',
 //   baseUrl,
-//   copyright: 'Copyright (c) 2018-present, 代码收容所',
-//   description: '天道酬勤，恒以致遠（大前端相关技术分享）',
+//   copyright: 'Copyright (c) 2018-present, 粥里有勺糖',
+//   description: '你的指尖,拥有改变世界的力量（大前端相关技术分享）',
 //   language: 'zh-cn',
 //   image: 'https://img.cdn.sugarat.top/mdImg/MTY3NDk5NTE2NzAzMA==674995167030',
-//   favicon: 'https://www.dmsrs.org/favicon.ico',
+//   favicon: 'https://sugarat.top/favicon.ico',
 // }
 
 // 所有配置项，详见文档: https://theme.sugarat.top/
@@ -22,14 +22,16 @@ const blogTheme = getThemeConfig({
 
   // 搜索
   // 默认开启pagefind离线的全文搜索支持（如使用其它的可以设置为false）
-  // 如果npx pagefind 时间过长，可以手动将其安装为项目依赖 pnpm add pagefind
   // search: false,
+
+  // markdown 图表支持（会增加一定的构建耗时）
+  // mermaid: true
 
   // 页脚
   footer: {
     // message 字段支持配置为HTML内容，配置多条可以配置为数组
     // message: '下面 的内容和图标都是可以修改的噢（当然本条内容也是可以隐藏的）',
-    copyright: 'MIT License | 代码收容所',
+    copyright: 'MIT License | 粥里有勺糖',
     // icpRecord: {
     //   name: '蜀ICP备19011724号',
     //   link: 'https://beian.miit.gov.cn/'
@@ -44,16 +46,16 @@ const blogTheme = getThemeConfig({
   themeColor: 'el-blue',
 
   // 文章默认作者
-  author: '代码收容所',
+  author: '粥里有勺糖',
 
   // 友链
   friend: [
     {
-      nickname: '代码收容所',
-      des: '天道酬勤，恒以致遠',
+      nickname: '粥里有勺糖',
+      des: '你的指尖用于改变世界的力量',
       avatar:
         'https://img.cdn.sugarat.top/mdImg/MTY3NDk5NTE2NzAzMA==674995167030',
-      url: 'https://www.dmsrs.org',
+      url: 'https://sugarat.top',
     },
     {
       nickname: 'Vitepress',
@@ -85,7 +87,7 @@ const blogTheme = getThemeConfig({
       {
         type: 'button',
         content: '作者博客',
-        link: 'https://www.dmsrs.org'
+        link: 'https://sugarat.top'
       },
       {
         type: 'button',