+++ packages/theme/src/utils/node/theme.ts
@@ -3,10 +3,11 @@ import fs from 'node:fs'
 import path from 'node:path'
 import process from 'node:process'
 import glob from 'fast-glob'
-import matter from 'gray-matter'
+import { getDefaultTitle, getFileLastModifyTime, getTextSummary, grayMatter, normalizePath } from '@sugarat/theme-shared'
+import type { SiteConfig } from 'vitepress'
 import type { Theme } from '../../composables/config/index'
 import { formatDate } from '../client'
-import { getDefaultTitle, getFileBirthTime, getFirstImagURLFromMD, getTextSummary } from './index'
+import { getFirstImagURLFromMD } from './index'
 
 export function patchDefaultThemeSideBar(cfg?: Partial<Theme.BlogConfig>) {
   return cfg?.blog !== false && cfg?.recommend !== false
@@ -22,37 +23,16 @@ export function patchDefaultThemeSideBar(cfg?: Partial<Theme.BlogConfig>) {
 }
 
 export function getPageRoute(filepath: string, srcDir: string) {
-  let route = filepath.replace('.md', '')
-  // 去除 srcDir 处理目录名
-  // TODO：优化 路径处理，同VitePress 内部一致
-  if (route.startsWith('./')) {
-    route = route.replace(
-      new RegExp(
-        `^\\.\\/${path
-          .join(srcDir, '/')
-          .replace(new RegExp(`\\${path.sep}`, 'g'), '/')}`
-      ),
-      ''
-    )
-  }
-  else {
-    route = route.replace(
-      new RegExp(
-        `^${path
-          .join(srcDir, '/')
-          .replace(new RegExp(`\\${path.sep}`, 'g'), '/')}`
-      ),
-      ''
-    )
-  }
+  const route = normalizePath(path.relative(srcDir, filepath))
+    .replace(/\.md$/, '')
   return `/${route}`
 }
 
 const defaultTimeZoneOffset = new Date().getTimezoneOffset() / -60
-export function getArticleMeta(filepath: string, route: string, timeZone = defaultTimeZoneOffset) {
-  const fileContent = fs.readFileSync(filepath, 'utf-8')
+export async function getArticleMeta(filepath: string, route: string, timeZone = defaultTimeZoneOffset) {
+  const fileContent = await fs.promises.readFile(filepath, 'utf-8')
 
-  const { data: frontmatter, excerpt, content } = matter(fileContent, {
+  const { data: frontmatter, excerpt, content } = grayMatter(fileContent, {
     excerpt: true,
   })
 
@@ -63,14 +43,13 @@ export function getArticleMeta(filepath: string, route: string, timeZone = defau
   if (!meta.title) {
     meta.title = getDefaultTitle(content)
   }
-  if (!meta.date) {
-    meta.date = formatDate(getFileBirthTime(filepath))
-  }
-  else {
-    meta.date = formatDate(
-      new Date(`${new Date(meta.date).toUTCString()}+${timeZone}`)
-    )
-  }
+  const date = await (
+    (meta.date
+       && new Date(`${new Date(meta.date).toUTCString()}+${timeZone}`))
+     || getFileLastModifyTime(filepath)
+  )
+  // 无法获取时兜底当前时间
+  meta.date = formatDate(date || new Date())
 
   // 处理tags和categories,兼容历史文章
   meta.categories
@@ -101,22 +80,42 @@ export function getArticleMeta(filepath: string, route: string, timeZone = defau
   }
   return meta as Theme.PageMeta
 }
-export function getArticles(cfg?: Partial<Theme.BlogConfig>) {
-  const srcDir = cfg?.srcDir || process.argv.slice(2)?.[1] || '.'
-  const files = glob.sync(`${srcDir}/**/*.md`, { ignore: ['node_modules'] })
+export async function getArticles(cfg: Partial<Theme.BlogConfig>, vpConfig: SiteConfig) {
+  const srcDir
+  = cfg?.srcDir || vpConfig.srcDir.replace(vpConfig.root, '').replace(/^\//, '')
+  || process.argv.slice(2)?.[1]
+  || '.'
+  const files = glob.sync(`${srcDir}/**/*.md`, { ignore: ['node_modules'], absolute: true })
 
-  // 文章数据
-  const pageData = files
-    .map((filepath) => {
-      const route = getPageRoute(filepath, srcDir)
-      const meta = getArticleMeta(filepath, route, cfg?.timeZone)
-      return {
-        route,
-        meta
-      }
+  const metaResults = files.reduce((prev, curr) => {
+    const route = getPageRoute(curr, vpConfig.srcDir)
+    const metaPromise = getArticleMeta(curr, route, cfg?.timeZone)
+
+    // 提前获取，有缓存取缓存
+    prev[curr] = {
+      route,
+      metaPromise
+    }
+    return prev
+  }, {} as Record<string, {
+    route: string
+    metaPromise: Promise<Theme.PageMeta>
+  }>)
+
+  const pageData: Theme.PageData[] = []
+
+  for (const file of files) {
+    const { route, metaPromise } = metaResults[file]
+    const meta = await metaPromise
+    if (meta.layout === 'home') {
+      continue
+    }
+    pageData.push({
+      route,
+      meta
     })
-    .filter(v => v.meta.layout !== 'home')
-  return pageData as Theme.PageData[]
+  }
+  return pageData
 }
 
 export function patchVPConfig(vpConfig: any, cfg?: Partial<Theme.BlogConfig>) {