+++ packages/vitepress-plugin-rss/src/type.ts
@@ -9,7 +9,7 @@ export type RSSOptions = Omit<FeedOptions, 'id'> & {
   filter?: (value: PostInfo, index: number, array: PostInfo[]) => boolean
   /**
    * 你的站点地址
-   * @example 'https://www.dmsrs.org'
+   * @example 'https://sugarat.top'
    */
   baseUrl: string
   /**